Refactor: Move two components to the popup folder.
