+++ src/app/app.module.ts
@@ -20,12 +20,12 @@ import {AppComponent} from './app.component';
 import {UnitDetailsComponent} from './unit-details/unit-details.component';
 import {CalculatorComponent} from './calculator/calculator.component';
 import {EquipmentDetailsComponent} from './equipment-details/equipment-details.component';
-import {UnitSelectionComponent} from './unit-selection/unit-selection.component';
+import {UnitSelectionComponent} from './popup/unit-selection/unit-selection.component';
 import {CalculationDisplayComponent} from './calculation-display/calculation-display.component';
 import {DatabaseClientService} from '../core/services/database-client.service';
 import {UnitsService} from '../core/services/units.service';
 import {UnitDetailsStatSumComponent} from './unit-details-stat-sum/unit-details-stat-sum.component';
-import {EquipmentSelectionComponent} from './equipment-selection/equipment-selection.component';
+import {EquipmentSelectionComponent} from './popup/equipment-selection/equipment-selection.component';
 import {EquipmentsDisplayComponent} from './equipments-display/equipments-display.component';
 import {SkillsDisplayComponent} from './skills-display/skills-display.component';
 import {CalculationDefensiveComponent} from './calculation-defensive/calculation-defensive.component';
