+++ src/app/equipments-display/equipments-display.component.ts
@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, OnDestroy, OnInit, Output} from '@angular/core';
 import {EquipmentSet} from '../../core/model/equipment-set.model';
 import {MatDialog} from '@angular/material';
-import {EquipmentSelectionComponent} from '../equipment-selection/equipment-selection.component';
+import {EquipmentSelectionComponent} from '../popup/equipment-selection/equipment-selection.component';
 import {Equipment} from '../../core/model/equipment.model';
 import {UnitsService} from '../../core/services/units.service';
 import {ISubscription} from 'rxjs/Subscription';
