+++ src/app/popup/equipment-selection/equipment-selection.component.css
