+++ src/app/popup/unit-selection/unit-selection.component.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 import {MatDialogRef} from '@angular/material';
-import {Unit} from '../../core/model/unit.model';
-import {UnitsService} from '../../core/services/units.service';
+import {Unit} from '../../../core/model/unit.model';
+import {UnitsService} from '../../../core/services/units.service';
 
 @Component({
   templateUrl: './unit-selection.component.html',
